Restructuring of repository, more documentation
